Extract session handlers
