+++ lib/ShortDev.Microsoft.ConnectedDevices.NearShare/Apps/NearShareApp.cs
@@ -67,7 +67,7 @@ internal sealed class NearShareApp(ConnectedDevicesPlatform cdp) : CdpAppBase(cd
                     }
                     _fileTransferToken = new()
                     {
-                        DeviceName = Channel.Session.DeviceName,
+                        DeviceName = Channel.Session.DeviceInfo?.Name ?? "UNKNOWM",
                         TotalBytes = bytesToSend,
                         Files = files
                     };
@@ -88,7 +88,7 @@ internal sealed class NearShareApp(ConnectedDevicesPlatform cdp) : CdpAppBase(cd
 
                     NearShareReceiver.OnReceivedUri(new()
                     {
-                        DeviceName = Channel.Session.DeviceName,
+                        DeviceName = Channel.Session.DeviceInfo?.Name ?? "UNKNOWM",
                         Uri = uri
                     });
